much better structured output + utils for generating pydantic schema
