+++ browser_use/controller/tests/test_controller_structured_output.py
@@ -3,17 +3,17 @@ import asyncio
 from pydantic import BaseModel
 
 from browser_use.agent.service import Agent
+from browser_use.agent.utils import create_pydantic_model_from_schema
 from browser_use.agent.views import AgentOutput
 from browser_use.controller.service import Controller
 from browser_use.llm.openai.chat import ChatOpenAI
-from browser_use.llm.schema import SchemaOptimizer
 
 
 class OutputModel(BaseModel):
 	"""Test output model"""
 
-	name: str
-	age: int
+	city: str
+	country: str
 
 
 async def test_optimized_schema():
@@ -26,20 +26,29 @@ async def test_optimized_schema():
 	# Create the agent output model with custom actions
 	agent_output_model = AgentOutput.type_with_custom_actions(ActionModel)
 
-	# Get original schema for comparison
-	original_schema = agent_output_model.model_json_schema()
+	# # Get original schema for comparison
+	# original_schema = agent_output_model.model_json_schema()
 
-	# Create the optimized schema
-	optimized_schema = SchemaOptimizer.create_optimized_json_schema(agent_output_model)
+	# # Create the optimized schema
+	# optimized_schema = SchemaOptimizer.create_optimized_json_schema(agent_output_model)
+
+	scgena = create_pydantic_model_from_schema(OutputModel.model_json_schema(), 'OutputModel')
 
 	agent = Agent(
-		task='What is the capital of France?',
-		llm=ChatOpenAI(model='gpt-4o'),
+		task='What is the capital of France? Do not use the internet, just output the done function.',
+		llm=ChatOpenAI(model='gpt-4.1-mini'),
 		controller=controller,
+		output_model_schema=scgena,
 	)
 
-	output = await agent.run()
-	x = output.structured_output
+	history = await agent.run()
+
+	if history.structured_output:
+		# print(history.structured_output.city, history.structured_output.country)
+		print(OutputModel.model_validate_json(history.final_result() or '{}'))
+	else:
+		print('No structured output')
+		print(history.final_result())
 
 
 if __name__ == '__main__':
