+++ pyproject.toml
@@ -32,6 +32,7 @@ dependencies = [
     "typing-extensions>=4.12.2",
     "uuid7>=0.1.0",
     "authlib>=1.6.0",
+    "datamodel-code-generator>=0.26.0",
     "google-genai>=1.21.1",
     "openai>=1.81.0",
     "anthropic>=0.54.0",
@@ -70,7 +71,6 @@ eval = [
     "anyio>=4.9.0",
     "Pillow>=11.2.1",
     "psutil>=7.0.0",
-    "datamodel-code-generator>=0.26.0",
 ]
 all = [
     "browser-use[cli,examples,aws]",
